+++ max_div/solver/_presets/_enum.py
@@ -2,4 +2,6 @@ from enum import StrEnum
 
 
 class SolverPreset(StrEnum):
-    DEFAULT = "default"
+    DEFAULT = "default"  # currently an alias for SMART
+    GUIDED = "guided"
+    SMART = "smart"