+++ max_div/solver/_presets/_main.py
@@ -4,7 +4,8 @@ from max_div.solver._solver_step import OptimizationStep
 from max_div.solver._strategies import InitializationStrategy
 
 from ._enum import SolverPreset
-from .preset_default import get_preset_strategies_default
+from .preset_guided import get_preset_strategies_guided
+from .preset_smart import get_preset_strategies_smart
 
 
 # =================================================================================================
@@ -18,8 +19,15 @@ def get_preset_strategies(
     hardware_speed_correction: float = 1.0,
 ) -> tuple[InitializationStrategy, list[OptimizationStep]]:
     match preset:
-        case SolverPreset.DEFAULT:
-            return get_preset_strategies_default(
+        case SolverPreset.GUIDED:
+            return get_preset_strategies_guided(
+                problem,
+                target_duration,
+                initialization_included,
+                hardware_speed_correction,
+            )
+        case SolverPreset.SMART | SolverPreset.DEFAULT:
+            return get_preset_strategies_smart(
                 problem,
                 target_duration,
                 initialization_included,